// components/calendar/interfaces.ts
import type { TEventColor } from "@/components/calendar/types";

export interface IUser {
  id: string;
  name: string;
  avatar: string | null;
}

export interface IEvent {
  id: string;
  startDate: string;
  endDate: string;
  title: string;
  color: TEventColor;
  description: string;
  user: IUser;
}

export interface ICalendarCell {
  day: number;
  currentMonth: boolean;
  date: Date;
}